backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
alembic>=1.13.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0
openpyxl>=3.1.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
feedparser>=6.0.0
rapidfuzz>=3.0.0
jinja2>=3.1.0
aiofiles>=24.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.9
weasyprint>=62.0
